requirements.txt
flask
pillow
pytesseract
virtualenv
